+++ cypress/integration/new-api/check-validate-if-logic.spec.js
@@ -2,6 +2,7 @@
 describe('Check ValidateIf logic', () => {
   beforeEach(() => {
     cy.visit('http://localhost:3000');
+    cy.wait(600);
   });
 
   it('load default data and clear field with ValidateIf logic', () => {
@@ -36,6 +37,16 @@ describe('Check ValidateIf logic', () => {
       'have.text',
       `{
   "name": "dep name 1",
+  "groups": [
+    {
+      "id": 33,
+      "name": "group name 1"
+    },
+    {
+      "id": 44,
+      "name": "group name 2"
+    }
+  ],
   "company": {
     "name": "comp name 1",
     "country": "USA",
@@ -59,14 +70,14 @@ describe('Check ValidateIf logic', () => {
     {
       "type": "car",
       "transportNumber": "DMC-12",
-      "numberOfWheels": 4,
-      "numberOfTurbines": ""
+      "numberOfTurbines": null,
+      "numberOfWheels": 4
     },
     {
       "type": "airplane",
       "transportNumber": "Boeing 737",
       "numberOfTurbines": 2,
-      "numberOfWheels": ""
+      "numberOfWheels": null
     }
   ]
 }`
@@ -138,6 +149,16 @@ describe('Check ValidateIf logic', () => {
       'have.text',
       `{
   "name": "dep name 1",
+  "groups": [
+    {
+      "id": 33,
+      "name": "group name 1"
+    },
+    {
+      "id": 44,
+      "name": "group name 2"
+    }
+  ],
   "company": {
     "name": "comp name 1",
     "country": "",
@@ -161,14 +182,14 @@ describe('Check ValidateIf logic', () => {
     {
       "type": "car",
       "transportNumber": "DMC-12",
-      "numberOfWheels": 4,
-      "numberOfTurbines": ""
+      "numberOfTurbines": null,
+      "numberOfWheels": 4
     },
     {
       "type": "airplane",
       "transportNumber": "Boeing 737",
       "numberOfTurbines": 2,
-      "numberOfWheels": ""
+      "numberOfWheels": null
     }
   ]
 }`
@@ -224,6 +245,16 @@ describe('Check ValidateIf logic', () => {
       'have.text',
       `{
   "name": "dep name 1",
+  "groups": [
+    {
+      "id": 33,
+      "name": "group name 1"
+    },
+    {
+      "id": 44,
+      "name": "group name 2"
+    }
+  ],
   "company": {
     "name": "comp name 1",
     "country": "USA",
@@ -247,14 +278,14 @@ describe('Check ValidateIf logic', () => {
     {
       "type": "car",
       "transportNumber": "DMC-12",
-      "numberOfWheels": 4,
-      "numberOfTurbines": ""
+      "numberOfTurbines": null,
+      "numberOfWheels": 4
     },
     {
       "type": "airplane",
       "transportNumber": "Boeing 737",
       "numberOfTurbines": 2,
-      "numberOfWheels": ""
+      "numberOfWheels": null
     }
   ]
 }`
@@ -271,6 +302,16 @@ describe('Check ValidateIf logic', () => {
       'have.text',
       `{
   "name": "dep name 1",
+  "groups": [
+    {
+      "id": 33,
+      "name": "group name 1"
+    },
+    {
+      "id": 44,
+      "name": "group name 2"
+    }
+  ],
   "company": {
     "name": "comp name 1",
     "country": "USA",
@@ -294,14 +335,14 @@ describe('Check ValidateIf logic', () => {
     {
       "type": "car",
       "transportNumber": "DMC-12",
-      "numberOfWheels": 4,
-      "numberOfTurbines": ""
+      "numberOfTurbines": null,
+      "numberOfWheels": 4
     },
     {
       "type": "airplane",
       "transportNumber": "Boeing 737",
       "numberOfTurbines": 2,
-      "numberOfWheels": ""
+      "numberOfWheels": null
     }
   ]
 }`