+++ libs/ngx-dynamic-form-builder/src/lib/lib.ts
@@ -5,11 +5,11 @@ import {
   FormControl,
   FormGroup,
   ValidationErrors,
+  ValidatorFn,
 } from '@angular/forms';
 import {
   ClassConstructor,
   classToPlain,
-  ClassTransformOptions,
   plainToClass,
 } from 'class-transformer-global-storage';
 import {
@@ -21,21 +21,13 @@ import cloneDeep from 'lodash.clonedeep';
 import lodashGet from 'lodash.get';
 import mergeWith from 'lodash.mergewith';
 import lodashSet from 'lodash.set';
-import {
-  BehaviorSubject,
-  catchError,
-  debounceTime,
-  from,
-  map,
-  Observable,
-  of,
-  ReplaySubject,
-  tap,
-} from 'rxjs';
+import { BehaviorSubject, from, Observable, of, ReplaySubject } from 'rxjs';
+import { catchError, debounceTime, delay, map, tap } from 'rxjs/operators';
 import { DEFAULT_CLASS_TRANSFORM_OPTIONS } from './constants/constants';
 import {
   ClassValidatorErrors,
   DeepPartial,
+  DynamicClassTransformOptions,
   DynamicFormArray,
   DynamicFormBuilderOptions,
   DynamicFormControl,
@@ -50,6 +42,7 @@ import {
   getGlobalDynamicFormBuilderOptionsSubject,
 } from './utils/global-dynamic-form-builder-options';
 import {
+  collectDynamicFormGroupErrors,
   isPrimitiveClass,
   isPrimitiveType,
   mergeErrors,
@@ -76,7 +69,7 @@ export function createFormControls<T = Record<string, unknown>>({
   rootFormGroup?: DynamicFormGroup<any>;
   metadata?: IDynamicControlMetadata;
   defaultValue: DeepPartial<T>;
-  dynamicFormBuilderOptions?: DynamicFormBuilderOptions;
+  dynamicFormBuilderOptions?: DynamicFormBuilderOptions<T>;
 }): DynamicFormGroup<T> {
   const isRoot = !rootFormGroup;
 
@@ -94,11 +87,16 @@ export function createFormControls<T = Record<string, unknown>>({
     });
   }
 
-  setupClassTransformMetadata<T>({
-    classType,
-    dynamicForm,
-    defaultMetadata: metadata,
-  });
+  if (dynamicFormBuilderOptions || rootFormGroup?.dynamicFormBuilderOptions) {
+    setupClassTransformMetadata<T>({
+      classType,
+      dynamicForm,
+      defaultMetadata: metadata,
+      dynamicFormBuilderOptions:
+        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+        dynamicFormBuilderOptions || rootFormGroup!.dynamicFormBuilderOptions,
+    });
+  }
 
   createAllFormGroupChildrenControls<T>({
     dynamicForm,
@@ -192,9 +190,14 @@ export function setValuesForControls<T = Record<string, unknown>>(
             ) {
               try {
                 control.patchValue(
-                  value !== undefined
+                  value !== undefined &&
+                    Object.getOwnPropertyDescriptor(
+                      value || {},
+                      metadataItem.propertyName
+                    )
                     ? (value as never)[metadataItem.propertyName]
-                    : ''
+                    : null,
+                  { emitEvent: false }
                 );
               } catch (err) {
                 console.log({ err, metadataItem, form });
@@ -217,12 +220,17 @@ export function setValuesForControls<T = Record<string, unknown>>(
           } else {
             const arrayControl = form.controls[
               metadataItem.propertyName
-            ] as FormArray;
-            let formArrayLength = arrayControl.length;
+            ] as DynamicFormArray;
 
-            while (formArrayLength !== 0) {
-              arrayControl.removeAt(0);
-              formArrayLength--;
+            if (
+              arrayControl?.classTransformMetadata?.isArray &&
+              arrayControl?.controls.length > 0
+            ) {
+              let formArrayLength = arrayControl.controls.length;
+              while (formArrayLength !== 0) {
+                arrayControl.removeAt(0);
+                formArrayLength--;
+              }
             }
 
             if (
@@ -230,9 +238,13 @@ export function setValuesForControls<T = Record<string, unknown>>(
               isPrimitiveClass(metadataItem.classType)
             ) {
               try {
-                (value !== undefined
-                  ? (value as never)[metadataItem.propertyName] || []
-                  : []
+                (
+                  (Object.getOwnPropertyDescriptor(
+                    value || {},
+                    metadataItem.propertyName
+                  )
+                    ? (value as never)[metadataItem.propertyName]
+                    : null) || []
                 ).forEach((item: never) => {
                   if (metadataItem.propertyName) {
                     arrayControl.push(formBuilder.control(item));
@@ -243,9 +255,13 @@ export function setValuesForControls<T = Record<string, unknown>>(
               }
             } else {
               try {
-                (value !== undefined
-                  ? (value as never)[metadataItem.propertyName] || []
-                  : []
+                (
+                  (Object.getOwnPropertyDescriptor(
+                    value || {},
+                    metadataItem.propertyName
+                  )
+                    ? (value as never)[metadataItem.propertyName]
+                    : null) || []
                 )
                   // eslint-disable-next-line @typescript-eslint/no-explicit-any
                   .forEach((item: any) => {
@@ -257,12 +273,25 @@ export function setValuesForControls<T = Record<string, unknown>>(
                           prop.propertyName &&
                           !itemKeys.includes(prop.propertyName)
                         ) {
-                          item[prop.propertyName] = '';
+                          item[prop.propertyName] = null;
                         }
                       });
                     }
                     if (metadataItem.propertyName) {
-                      arrayControl.push(formBuilder.group(item));
+                      const control = createFormControls({
+                        classType: metadataItem.classType,
+                        formBuilder,
+                        metadata: metadataItem,
+                        defaultValue: item,
+                        rootFormGroup: form.root as DynamicFormGroup<
+                          unknown,
+                          unknown
+                        >,
+                        dynamicFormBuilderOptions: (
+                          form.root as DynamicFormGroup<unknown, unknown>
+                        ).dynamicFormBuilderOptions,
+                      });
+                      arrayControl.push(control);
                     }
                   });
               } catch (err) {
@@ -282,7 +311,7 @@ export function setValidatorsToControls<T>(
   rootFormGroup?: DynamicFormGroup<T>
 ) {
   if (!rootFormGroup) {
-    rootFormGroup = form;
+    rootFormGroup = form.root as DynamicFormGroup<T>;
   }
 
   addCommonAsyncValidatorToRootForm(rootFormGroup);
@@ -314,7 +343,49 @@ export function setValidatorsToControls<T>(
               ) {
                 form.controls[metadataItem.propertyName].setAsyncValidators(
                   rootFormGroup.commonAsyncValidator
+                ); /*
+                form.controls[
+                  metadataItem.propertyName
+                ].updateValueAndValidity();*/
+              }
+
+              if (rootFormGroup.dynamicFormBuilderOptions.angularValidators) {
+                const { angularFormValidatorPath } = findControlPaths(
+                  form.controls[metadataItem.propertyName]
                 );
+                const angularValidators = lodashGet(
+                  rootFormGroup.dynamicFormBuilderOptions.angularValidators,
+                  angularFormValidatorPath
+                  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+                );
+                if (angularValidators) {
+                  let asyncValidators = [];
+                  if (angularValidators.asyncValidators) {
+                    asyncValidators = angularValidators.asyncValidators;
+                    if (typeof asyncValidators === 'function') {
+                      asyncValidators = [asyncValidators];
+                    }
+                  }
+                  let validators =
+                    angularValidators?.validators || angularValidators;
+                  if (typeof validators === 'function') {
+                    validators = [validators];
+                  }
+                  form.controls[metadataItem.propertyName].setAsyncValidators([
+                    ...asyncValidators,
+                    ...validators
+                      .filter(Boolean)
+                      .map(
+                        (validator: ValidatorFn) =>
+                          (control: AbstractControl) =>
+                            Promise.resolve(validator(control))
+                      ),
+                    rootFormGroup.commonAsyncValidator,
+                  ]); /*
+                  form.controls[
+                    metadataItem.propertyName
+                  ].updateValueAndValidity();*/
+                }
               }
             }
           } catch (err) {
@@ -340,9 +411,13 @@ export function setValidatorsToControls<T>(
             throw new Error('commonAsyncValidator not set');
           }
           try {
-            (
+            const controls = (
               form.controls[metadataItem.propertyName] as DynamicFormArray
-            ).controls.forEach((control, propertyIndex) => {
+            ).controls;
+            (!Array.isArray(controls) && metadataItem.isArray
+              ? []
+              : controls
+            ).forEach((control, propertyIndex) => {
               (control as DynamicFormGroup).classTransformMetadata = {
                 ...metadataItem,
                 propertyIndex,
@@ -355,6 +430,7 @@ export function setValidatorsToControls<T>(
                 !control.hasAsyncValidator(rootFormGroup.commonAsyncValidator)
               ) {
                 control.setAsyncValidators(rootFormGroup.commonAsyncValidator);
+                /* control.updateValueAndValidity(); */
               }
             });
             if (!rootFormGroup) {
@@ -368,15 +444,22 @@ export function setValidatorsToControls<T>(
               (
                 form.controls[metadataItem.propertyName] as DynamicFormArray
               ).setAsyncValidators(rootFormGroup.commonAsyncValidator);
+              /* (
+                form.controls[metadataItem.propertyName] as DynamicFormArray
+              ).updateValueAndValidity(); */
             }
           } catch (err) {
             console.log({ err, metadataItem, form });
           }
         } else {
           try {
-            (
+            const controls = (
               form.controls[metadataItem.propertyName] as DynamicFormArray
-            ).controls.forEach((control, propertyIndex) => {
+            ).controls;
+            (!Array.isArray(controls) && metadataItem.isArray
+              ? []
+              : controls
+            ).forEach((control, propertyIndex) => {
               (control as DynamicFormGroup).classTransformMetadata = {
                 ...metadataItem,
                 propertyIndex,
@@ -428,7 +511,7 @@ export function setValidatorsToControls<T>(
                           );
                         }
                       } else {
-                        controlValue[prop.propertyName] = '';
+                        controlValue[prop.propertyName] = null;
                         if (
                           // eslint-disable-next-line @typescript-eslint/no-explicit-any
                           !(control as DynamicFormGroup<any, any>).controls[
@@ -464,6 +547,7 @@ export function setValidatorsToControls<T>(
                 !control.hasAsyncValidator(rootFormGroup.commonAsyncValidator)
               ) {
                 control.setAsyncValidators(rootFormGroup.commonAsyncValidator);
+                /* control.updateValueAndValidity(); */
               }
             });
             if (!rootFormGroup) {
@@ -477,6 +561,9 @@ export function setValidatorsToControls<T>(
               (
                 form.controls[metadataItem.propertyName] as DynamicFormArray
               ).setAsyncValidators(rootFormGroup.commonAsyncValidator);
+              /* (
+                form.controls[metadataItem.propertyName] as DynamicFormArray
+              ).updateValueAndValidity(); */
             }
           } catch (err) {
             console.log({ err, metadataItem, form });
@@ -522,17 +609,22 @@ export function validateAllFormFields(form: FormGroup) {
   });
 }
 
+// custom: getMetadata args have been changed. Before optional parameters, these were added:
+// dynamicFormBuilderOptions
+// currentDepth
+// propertyPath
 function getMetadata(
   classType: ClassConstructor<unknown>,
-  dynamicFormBuilderOptions: DynamicFormBuilderOptions,
+  dynamicFormBuilderOptions: DynamicFormBuilderOptions<any>,
   currentDepth:number,
   currentPath:string[],
-  classTransformOptions?: ClassTransformOptions,
+  classTransformOptions?: DynamicClassTransformOptions,
+  isRoot?:boolean,
 ): IDynamicControlMetadata {
-  const classTransformerMetadataStorage =
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-   getGlobal()['classTransformerMetadataStorage'] || undefined;
-  if (!classTransformerMetadataStorage) {
+	const classTransformerMetadataStorage =
+	  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+	 getGlobal()['classTransformerMetadataStorage'] || undefined;
+	if (!classTransformerMetadataStorage) {
     throw new Error(
       'classTransformerMetadataStorage not set in windows, please use the "class-transformer-global-storage" instead of "class-transformer"'
     );
@@ -556,7 +648,7 @@ function getMetadata(
   
 
   // need for create all link for multi types
-  if (classTransformOptions) {
+  if (isRoot) {
     while (
       !prevMultiTypes ||
       !multiTypes ||
@@ -569,9 +661,13 @@ function getMetadata(
         .map((meta: any) =>
           Array.from(meta[1])
             // eslint-disable-next-line @typescript-eslint/no-explicit-any
-            .filter((sub: any) => sub[1].options.discriminator)
+            .filter((sub: any) => {
+              return sub[1].options.discriminator;
+            })
             // eslint-disable-next-line @typescript-eslint/no-explicit-any
-            .map((sub: any) => ({ ...sub[1], ...sub[1].options.discriminator }))
+            .map((sub: any) => {
+              return { ...sub[1], ...sub[1].options.discriminator };
+            })
         )
         .filter((arr) => arr.length > 0)
         .map((arr) =>
@@ -584,7 +680,7 @@ function getMetadata(
         )
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
         .reduce((all: any, cur: any) => [...all, ...cur], []);
-      multiTypes.forEach((multiType) =>
+      (multiTypes || []).forEach((multiType) =>
         multiType.types.forEach((type: { name: string }) =>
           plainToClass(
             multiType.classType,
@@ -612,7 +708,9 @@ function getMetadata(
         .map((meta: any) =>
           Array.from(meta[1])
             // eslint-disable-next-line @typescript-eslint/no-explicit-any
-            .filter((sub: any) => sub[1].options.discriminator)
+            .filter((sub: any) => {
+              return sub[1].options.discriminator;
+            })
             // eslint-disable-next-line @typescript-eslint/no-explicit-any
             .map((sub: any) => ({ ...sub[1], ...sub[1].options.discriminator }))
         )
@@ -627,7 +725,7 @@ function getMetadata(
         )
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
         .reduce((all: any, cur: any) => [...all, ...cur], []);
-      multiTypes.forEach((multiType) =>
+      (multiTypes || []).forEach((multiType) =>
         multiType.types.forEach((type: { name: string }) =>
           plainToClass(
             multiType.classType,
@@ -666,12 +764,20 @@ function getMetadata(
     ...(ancestorsProperties.length > 0 ? ancestorsProperties[0] : []),
   ].filter(
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (existsLimit: any, existsIndex: any, items: any) =>
-      existsIndex ===
-      items.findIndex(
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        (it: any) => it.propertyName === existsLimit.propertyName
-      )
+    (existsLimit: any, existsIndex: any, items: any) => {
+      return (
+        (classTransformOptions?.excludeGroups
+          ? !classTransformOptions?.excludeGroups.find((excludeGroup) =>
+              existsLimit?.options?.groups?.includes(excludeGroup)
+            )
+          : true) &&
+        existsIndex ===
+          items.findIndex(
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            (it: any) => it.propertyName === existsLimit.propertyName
+          )
+      );
+    }
   );
 
   const depthLimitExceeded = currentDepth >= (dynamicFormBuilderOptions.maxNestedModelDepth as number);
@@ -690,38 +796,56 @@ function getMetadata(
       );
 
 	  
+	  // custom: getMetadata args have been changed. Before optional parameters, these were added:
+	  // dynamicFormBuilderOptions
+	  // currentDepth
+	  // propertyPath
+
+	  // old implementation:
+	  /*
 	  if (propertyMetadata) {
+	    properties[index] = {
+	      ...getMetadata(
+	        propertyMetadata.typeFunction(),
+	        classTransformOptions,
+	        false
+	      ),
+	      isArray: Array === propertyMetadata.reflectedType,
+	      propertyName: exposeMetadataItem.propertyName,
+		}
+	  };
+	  */
+      if (propertyMetadata) {
+	    // custom: skip if depthLimitExceeded
 		if(depthLimitExceeded) {
-	      // prop exceeds limit, it should not be created at all, so remove it from list
 		  console.warn('ngx-dynamic-form-builder has hit depth limit!',propertyMetadata)
 		  delete properties[index];
 		} else {
 		  const propertyPath = [...currentPath,exposeMetadataItem.propertyName];
-		  // only include submodel type data if allowedNestedModels is not set
-		  // or, if it is defined, only allow if the property path is on list.
-		  if(!Array.isArray(dynamicFormBuilderOptions.allowedNestedModels))
-		  {
-		  	properties[index] = {
-		  	  ...getMetadata(propertyMetadata.typeFunction(),dynamicFormBuilderOptions,currentDepth+1,propertyPath,undefined),
-		  	  isArray: Array === propertyMetadata.reflectedType,
-		  	  propertyName: exposeMetadataItem.propertyName,
-		  	};
+
+		  const assignProperty = () => {
+			properties[index] = {
+				...getMetadata(
+					propertyMetadata.typeFunction(),
+					dynamicFormBuilderOptions,
+					currentDepth+1,
+					propertyPath,
+					undefined,
+					currentDepth===0
+				),
+				isArray: Array === propertyMetadata.reflectedType,
+				propertyName: exposeMetadataItem.propertyName,
+			  }
+		  }
+		  // custom: this is new logic. Instead of assigning all, 
+		  // metadata will be filtered out if allowedNestedModels are defined and do not include this prop.
+		  if(!Array.isArray(dynamicFormBuilderOptions.allowedNestedModels)) {
+			assignProperty();
 		  } else {
 		    if(dynamicFormBuilderOptions.allowedNestedModels.includes(propertyPath.join('.'))) {
-		  	  properties[index] = {
-		  		...getMetadata(propertyMetadata.typeFunction(),dynamicFormBuilderOptions,currentDepth+1,propertyPath,undefined),
-		  		isArray: Array === propertyMetadata.reflectedType,
-		  		propertyName: exposeMetadataItem.propertyName,
-		  	  };
+				assignProperty();
 		    } else {
-		      // prop should be excluded, add no metadata info at all
 		      delete properties[index];
-		      // properties[index] = {
-		      //   classType: null,
-		      //   properties: [],
-		      //   isArray: false,
-		      //   propertyName: exposeMetadataItem.propertyName,
-		      // };
 		    }
 		  }
 		}
@@ -737,7 +861,7 @@ function getMetadata(
     }
   );
 
-  return {
+  const metadata = {
     classType,
     properties:properties.filter(prop=>typeof prop !== 'undefined'), // as entries may have been deleted, we need to filter those out to not break other code
     withAncestors:
@@ -746,16 +870,20 @@ function getMetadata(
     isArray: false,
     propertyName: null,
   };
+
+  return metadata;
 }
 
 function setupClassTransformMetadata<T = Record<string, unknown>>({
   defaultMetadata,
   classType,
   dynamicForm,
+  dynamicFormBuilderOptions,
 }: {
   defaultMetadata: IDynamicControlMetadata | undefined;
   classType: ClassConstructor<T> | null;
   dynamicForm: DynamicFormGroup<T, T>;
+  dynamicFormBuilderOptions: DynamicFormBuilderOptions<T>;
 }) {
   if (!defaultMetadata && classType) {
     dynamicForm.classTransformMetadata = getMetadata(
@@ -763,6 +891,8 @@ function setupClassTransformMetadata<T = Record<string, unknown>>({
       dynamicForm.dynamicFormBuilderOptions,
 	  0,
 	  [],
+      dynamicFormBuilderOptions.classTransformOptions,
+      true,
     );
   } else {
     dynamicForm.classTransformMetadata =
@@ -778,7 +908,7 @@ function setupDynamicFormBuilderOptions<T = Record<string, unknown>>({
   dynamicFormBuilderOptions,
 }: {
   dynamicForm: DynamicFormGroup<T, T>;
-  dynamicFormBuilderOptions: DynamicFormBuilderOptions | undefined;
+  dynamicFormBuilderOptions: DynamicFormBuilderOptions<T> | undefined;
 }) {
   if (!dynamicForm) {
     throw new Error('dynamicForm not set');
@@ -816,7 +946,7 @@ function createAllFormGroupChildrenControls<T = Record<string, unknown>>({
   formBuilder: FormBuilder;
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   rootFormGroup: DynamicFormGroup<any, any> | undefined;
-  dynamicFormBuilderOptions: DynamicFormBuilderOptions | undefined;
+  dynamicFormBuilderOptions: DynamicFormBuilderOptions<T> | undefined;
 }) {
   dynamicForm.classTransformMetadata.properties.forEach(
     (metadataItem, index) => {
@@ -827,7 +957,8 @@ function createAllFormGroupChildrenControls<T = Record<string, unknown>>({
           Object.getOwnPropertyDescriptor(
             defaultValue || {},
             metadataItem.propertyName
-          )
+          ) ||
+          metadataItem.isArray
         ) {
           if (!metadataItem.isArray) {
             if (
@@ -896,7 +1027,7 @@ function addDynamicTypedObjectAsFormArray<T = Record<string, unknown>>({
   formBuilder: FormBuilder;
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   rootFormGroup: DynamicFormGroup<any, any> | undefined;
-  dynamicFormBuilderOptions: DynamicFormBuilderOptions | undefined;
+  dynamicFormBuilderOptions: DynamicFormBuilderOptions<T> | undefined;
   dynamicForm: DynamicFormGroup<T, T>;
   index: number;
 }) {
@@ -977,7 +1108,7 @@ function addDynamicTypedObjectAsFormGroup<T = Record<string, unknown>>({
   formBuilder: FormBuilder;
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   rootFormGroup: DynamicFormGroup<any, any> | undefined;
-  dynamicFormBuilderOptions: DynamicFormBuilderOptions | undefined;
+  dynamicFormBuilderOptions: DynamicFormBuilderOptions<T> | undefined;
   dynamicForm: DynamicFormGroup<T, T>;
   index: number;
 }) {
@@ -993,7 +1124,7 @@ function addDynamicTypedObjectAsFormGroup<T = Record<string, unknown>>({
           all[cur.propertyName] = [];
         } else {
           if (!cur.classType || isPrimitiveClass(cur.classType)) {
-            all[cur.propertyName] = '';
+            all[cur.propertyName] = null;
           } else {
             all[cur.propertyName] = {};
           }
@@ -1034,7 +1165,12 @@ function addDynamicPrimitiveObjectAsFormGroup<T = Record<string, unknown>>({
   if (!metadataItem.propertyName) {
     throw new Error('propertyName not set');
   }
-  const value = (defaultValue as never)[metadataItem.propertyName] || '';
+  const value = Object.getOwnPropertyDescriptor(
+    defaultValue,
+    metadataItem.propertyName
+  )
+    ? (defaultValue as never)[metadataItem.propertyName]
+    : undefined;
   const control = formBuilder.control(value) as DynamicFormControl;
   dynamicForm.addControl(metadataItem.propertyName, control);
   control.classTransformMetadata = metadataItem;
@@ -1057,7 +1193,7 @@ function addDynamicMethodsToRootFormGroup<T = Record<string, unknown>>({
   );
 
   dynamicForm.patchDynamicFormBuilderOptions = (
-    dynamicFormBuilderOptions: DynamicFormBuilderOptions
+    dynamicFormBuilderOptions: DynamicFormBuilderOptions<T>
   ) => {
     if (!dynamicForm) {
       throw new Error('rootFormGroup not set');
@@ -1067,7 +1203,7 @@ function addDynamicMethodsToRootFormGroup<T = Record<string, unknown>>({
       ...dynamicFormBuilderOptions,
     };
     setValidatorsToControls(formBuilder, dynamicForm);
-    dynamicForm.patchValue(dynamicForm.value);
+    dynamicForm.patchValue(dynamicForm.value, { emitEvent: false });
   };
 
   dynamicForm.getObject = () => {
@@ -1121,7 +1257,7 @@ function addDynamicMethodsToRootFormGroup<T = Record<string, unknown>>({
     dynamicForm.externalErrorsSubject.next(
       recursiveRemoveDynamicControlOptions(cloneDeep(externalErrors))
     );
-    dynamicForm.patchValue(dynamicForm.value);
+    dynamicForm.patchValue(dynamicForm.value, { emitEvent: false });
   };
 
   dynamicForm.getExternalErrors = () => {
@@ -1169,7 +1305,7 @@ function addDynamicMethodsToRootFormGroup<T = Record<string, unknown>>({
     },
   });
 
-  dynamicFormProxy.patchValue(dynamicFormProxy.value);
+  dynamicFormProxy.patchValue(dynamicFormProxy.value, { emitEvent: false });
   setValidatorsToControls(formBuilder, dynamicFormProxy, rootFormGroup);
   return dynamicFormProxy;
 }
@@ -1332,6 +1468,7 @@ function setDynamicControlValue<T>(
 function findControlPaths(
   control: DynamicFormProperties<unknown> | AbstractControl
 ) {
+  const angularFormValidatorPath: string[] = [];
   const inputPath: string[] = [];
   const errorsPath: string[] = [];
   let ctrl = control as DynamicFormProperties;
@@ -1340,17 +1477,26 @@ function findControlPaths(
       ctrl.classTransformMetadata?.propertyIndex?.toString() ||
       ctrl.classTransformMetadata?.propertyName ||
       null;
+    const parent = ctrl.parent as DynamicFormProperties;
     if (key !== null) {
       inputPath.push(key);
       errorsPath.push(key);
-      if (ctrl.parent?.parent) {
+      if (parent?.parent) {
         errorsPath.push('children');
       }
+      if (!parent?.classTransformMetadata?.isArray) {
+        angularFormValidatorPath.push(key);
+      }
     }
-    ctrl = ctrl.parent as DynamicFormProperties;
+    ctrl = parent;
   }
   const controlPath = inputPath.reverse().join('.');
-  return { inputPath, errorsPath, controlPath };
+  return {
+    inputPath,
+    errorsPath,
+    controlPath,
+    angularFormValidatorPath: angularFormValidatorPath.reverse().join('.'),
+  };
 }
 
 function setFormControlClassValidatorErrors<T>(
@@ -1423,8 +1569,11 @@ function setRootFormGroupClassValidatorErrors<T>(
       new BehaviorSubject<ShortValidationErrors>({});
   }
   rootFormGroup.customValidateErrors.next(
-    transformClassValidatorErrorsToShortValidationErrors(
-      rootFormGroup.classValidatorErrors
+    mergeErrors(
+      collectDynamicFormGroupErrors(rootFormGroup),
+      transformClassValidatorErrorsToShortValidationErrors(
+        rootFormGroup.classValidatorErrors
+      )
     )
   );
 }
@@ -1440,8 +1589,22 @@ function subscribeToRootFormGroupValueChanges<T>(
             if (!rootFormGroup) {
               throw Error('rootFormGroup not set');
             }
-
             rootFormGroup.commonAsyncValidatorFirstChanged = true;
+          }),
+          delay(100),
+          tap(() => {
+            if (!rootFormGroup) {
+              throw Error('rootFormGroup not set');
+            }
+
+            rootFormGroup.customValidateErrors.next(
+              mergeErrors(
+                collectDynamicFormGroupErrors(rootFormGroup),
+                transformClassValidatorErrorsToShortValidationErrors(
+                  rootFormGroup.classValidatorErrors
+                )
+              )
+            );
           }),
           catchError((err) => {
             console.log({ err, rootFormGroup });
@@ -1462,7 +1625,7 @@ function subscribeToGlobalDynamicFormBuilderOptionsChange<T>(
         .pipe(
           tap(() => {
             setValidatorsToControls(formBuilder, rootFormGroup);
-            rootFormGroup.patchValue(rootFormGroup.value);
+            rootFormGroup.patchValue(rootFormGroup.value, { emitEvent: false });
           }),
           catchError((err) => {
             console.log({ err, rootFormGroup });